Fix `IfLet` View
The `IfLet` view returned an `EmptyView` in the `else` case, which caused
a task modifier not to be called. This issue broke the TransducerView
in cases where the proxy was not provided, and a default proxy has been
created by the TransducerView.

This commit changes the `else` path such that it returns "poper" view,
aka a `Color.Clear` which fixed the issue.